Fix Tor scheduler bug
channel_num_cells_writeable was subtracting the outgoing_queue cells count from
the channeltls writeable count, probably an attempt to stop overcommitting
cells to the channel. However it's the scheduler's responsibility to write the
outgoing_queue cells out, and with the previous behaviour this resulted in the
scheduler trying to write much less than it should when it has outgoing_queued
cells.

With the new behaviour, we do try to write as much as the channel lower level
thinks it can take, but if the outgoing_queue doesn't fully flush we return
immediately without attempting new cell writes on the channel. This ensures the
outgoing_queue doesn't grow to unbounded lengths and results in fewer dynamic
allocations when the new cell writes would inevitably become queued.

Also remove the assumption that writing less than the lower level would accept
necessarily means all queues are empty. Now the scheduler asks that question
directly to the channel.